@@ -49,35 +49,35 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.1.2",
-    "@angular/cdk": "^14.1.1",
-    "@angular/common": "^14.1.2",
-    "@angular/compiler": "^14.1.2",
-    "@angular/core": "^14.1.2",
-    "@angular/forms": "^14.1.2",
-    "@angular/material": "^14.1.1",
-    "@angular/platform-browser": "^14.1.2",
-    "@angular/platform-browser-dynamic": "^14.1.2",
-    "@angular/router": "^14.1.2",
-    "@angular/service-worker": "^14.1.2",
+    "@angular/animations": "^14.3.0",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/router": "^14.3.0",
+    "@angular/service-worker": "^14.3.0",
     "@azure/msal-angular": "^2.4.1",
     "@azure/msal-browser": "^2.28.1",
-    "@igo2/auth": "1.14.0",
-    "@igo2/common": "1.14.0",
-    "@igo2/context": "1.14.0",
-    "@igo2/core": "1.14.0",
-    "@igo2/geo": "1.14.0",
-    "@igo2/integration": "1.14.0",
-    "@igo2/utils": "1.14.0",
+    "@igo2/auth": "1.15.1",
+    "@igo2/common": "1.15.1",
+    "@igo2/context": "1.15.1",
+    "@igo2/core": "1.15.1",
+    "@igo2/geo": "1.15.1",
+    "@igo2/integration": "1.15.1",
+    "@igo2/utils": "1.15.1",
     "@mat-datetimepicker/core": "~10.0.0",
     "@turf/buffer": "^6.5.0",
     "@turf/point-on-feature": "^6.5.0",
-    "ajv": "^8.11.0",
+    "ajv": "^6.9.1",
     "angular-shepherd": "14.0.0",
     "bowser": "^2.10.0",
     "core-js": "^3.24.1",
     "fast-text-encoding": "^1.0.4",
     "file-saver": "^2.0.2",
+    "flexsearch": "0.7.21",
     "hammerjs": "^2.0.8",
     "html2canvas": "^1.4.1",
     "jspdf": "^2.5.1",
@@ -136,13 +136,16 @@
     "karma-coverage": "^2.2.0",
     "karma-jasmine": "~4.0.2",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "ng-packagr": "^14.1.0",
+    "ng-packagr": "^14.2.2",
     "npx": "^10.2.2",
     "protractor": "~7.0.0",
     "sass": "^1.54.4",
     "ts-node": "^10.9.1",
     "typescript": "~4.6.4",
     "webdriver-manager": "^12.1.6",
-    "webpack": "^5.74.0"
+    "webpack": "^5.76.0"
+  },
+  "engines": {
+    "node": ">=16.19.0"
   }
 }
