@@ -7,8 +7,8 @@ import {
   ElementRef
 } from '@angular/core';
 import { ActivatedRoute, Params } from '@angular/router';
-import { Subscription, of, BehaviorSubject, combineLatest, zip } from 'rxjs';
-import { debounceTime, take, pairwise, skipWhile, first, concatMap, delay } from 'rxjs/operators';
+import { Subscription, of, BehaviorSubject, combineLatest } from 'rxjs';
+import { debounceTime, take, pairwise, skipWhile, first } from 'rxjs/operators';
 import { MatDialog, MatDialogConfig, MatDialogRef } from '@angular/material/dialog';
 import MapBrowserEvent from 'ol/MapBrowserEvent';
 import * as olProj from 'ol/proj';
@@ -101,7 +101,6 @@ import { WelcomeWindowService } from './welcome-window/welcome-window.service';
 import { MatPaginator } from '@angular/material/paginator';
 import { ObjectUtils } from '@igo2/utils';
 import olFormatGeoJSON from 'ol/format/GeoJSON';
-import { PwaService } from '../../services/pwa.service';
 
 @Component({
   selector: 'app-portal',
@@ -126,6 +125,8 @@ export class PortalComponent implements OnInit, OnDestroy {
   public hasHomeExtentButton = false;
   public showMenuButton = true;
   public showSearchBar = true;
+  public showOfflineButton = false;
+  public showWakeLockButton = false;
   public showRotationButtonIfNoRotation = false;
   public hasFeatureEmphasisOnSelection: Boolean = false;
   public workspaceNotAvailableMessage: String = 'workspace.disabled.resolution';
@@ -163,6 +164,7 @@ export class PortalComponent implements OnInit, OnDestroy {
   private sidenavMediaAndOrientation$$: Subscription;
 
   public igoSearchPointerSummaryEnabled: boolean;
+  public igoReverseSearchCoordsFormatEnabled: boolean;
 
   public toastPanelForExpansionOpened = true;
   private activeWidget$$: Subscription;
@@ -243,6 +245,10 @@ export class PortalComponent implements OnInit, OnDestroy {
     } else {
       this.map.viewController.setPadding({bottom: 0});
     }
+  }
+
+  get contextUri(): string {
+    return this.contextState.context$?.getValue() ? this.contextState.context$.getValue().uri : undefined;
   }
 
   get toastPanelShown(): boolean {
@@ -332,7 +338,6 @@ export class PortalComponent implements OnInit, OnDestroy {
     private storageService: StorageService,
     private editionWorkspaceService: EditionWorkspaceService,
     private directionState: DirectionState,
-    private pwaService: PwaService,
     private configFileToGeoDBService: ConfigFileToGeoDBService
   ) {
     this.hasExpansionPanel = this.configService.getConfig('hasExpansionPanel');
@@ -344,8 +349,13 @@ export class PortalComponent implements OnInit, OnDestroy {
       this.configService.getConfig('showRotationButtonIfNoRotation');
     this.showMenuButton = this.configService.getConfig('showMenuButton') === undefined ? true :
       this.configService.getConfig('showMenuButton');
-    this.showSearchBar = this.configService.getConfig('showSearchBar') === undefined ? true :
-      this.configService.getConfig('showSearchBar');
+    this.showSearchBar = this.configService.getConfig('searchBar.showSearchBar') === undefined ? true :
+      this.configService.getConfig('searchBar.showSearchBar');
+    this.showOfflineButton = this.configService.getConfig('offlineButton') === undefined ? false :
+      this.configService.getConfig('offlineButton');
+    this.showWakeLockButton = this.configService.getConfig('wakeLockApiButton') === undefined ? false :
+      this.configService.getConfig('wakeLockApiButton');
+
     this.forceCoordsNA = this.configService.getConfig('app.forceCoordsNA');
     this.hasFeatureEmphasisOnSelection = this.configService.getConfig('hasFeatureEmphasisOnSelection');
 
@@ -354,6 +364,8 @@ export class PortalComponent implements OnInit, OnDestroy {
     if (this.igoSearchPointerSummaryEnabled === undefined) {
       this.igoSearchPointerSummaryEnabled = this.storageService.get('searchPointerSummaryEnabled') as boolean || false;
     }
+
+    this.igoReverseSearchCoordsFormatEnabled = this.storageService.get('reverseSearchCoordsFormatEnabled') as boolean || false;
   }
 
   ngOnInit() {
@@ -503,63 +515,6 @@ export class PortalComponent implements OnInit, OnDestroy {
       if (configFileToGeoDBService) {
         this.configFileToGeoDBService.load(configFileToGeoDBService);
       }
-    }
-    // this.initSW();
-  }
-
-  private initSW() {
-    const dataDownload = this.configService.getConfig('pwa.dataDownload');
-    if ('serviceWorker' in navigator && dataDownload) {
-      let downloadMessage;
-      let currentVersion;
-      const dataLoadSource = this.storageService.get('dataLoadSource');
-      navigator.serviceWorker.ready.then((registration) => {
-        console.log('Service Worker Ready');
-        this.http.get('ngsw.json').pipe(
-          concatMap((ngsw: any) => {
-            const datas$ = [];
-            let hasDataInDataDir: boolean = false;
-            if (ngsw) {
-              // IF FILE NOT IN THIS LIST... DELETE?
-              currentVersion = ngsw.appData.version;
-              const cachedDataVersion = this.storageService.get('cachedDataVersion');
-              if (currentVersion !== cachedDataVersion && dataLoadSource === 'pending') {
-                this.pwaService.updates.checkForUpdate();
-              }
-              if (dataLoadSource === 'newVersion' || !dataLoadSource) {
-                ((ngsw as any).assetGroups as any).map((assetGroup) => {
-                  if (assetGroup.name === 'contexts') {
-                    const elemToDownload = assetGroup.urls.concat(assetGroup.files).filter(f => f);
-                    elemToDownload.map((url, i) => datas$.push(this.http.get(url).pipe(delay(750))));
-                  }
-                });
-                if (hasDataInDataDir) {
-                  const message = this.languageService.translate.instant('pwa.data-download-start');
-                  downloadMessage = this.messageService
-                    .info(message, undefined, { disableTimeOut: true, progressBar: false, closeButton: true, tapToDismiss: false });
-                  this.storageService.set('cachedDataVersion', currentVersion);
-                }
-                return zip(...datas$);
-              }
-
-            }
-            return zip(...datas$);
-          })
-        )
-          .pipe(delay(1000))
-          .subscribe(() => {
-            if (downloadMessage) {
-              this.messageService.remove((downloadMessage as any).toastId);
-              const message = this.languageService.translate.instant('pwa.data-download-completed');
-              this.messageService.success(message, undefined, { timeOut: 40000 });
-              if (currentVersion) {
-                this.storageService.set('dataLoadSource', 'pending');
-                this.storageService.set('cachedDataVersion', currentVersion);
-              }
-            }
-          });
-
-      });
     }
   }
 
@@ -749,7 +704,7 @@ export class PortalComponent implements OnInit, OnDestroy {
     if (this.routeParams?.search && term !== this.routeParams.search) {
       this.searchState.deactivateCustomFilterTermStrategy();
     }
-
+    this.searchBarTerm = term;
     this.searchState.setSearchTerm(term);
     const termWithoutHashtag = term.replace(/(#[^\s]*)/g, '').trim();
     if (termWithoutHashtag.length < 2) {
@@ -787,6 +742,11 @@ export class PortalComponent implements OnInit, OnDestroy {
 
   onSearchResultsGeometryStatusChange(value) {
     this.searchState.setSearchResultsGeometryStatus(value);
+  }
+
+  onReverseCoordsFormatStatusChange(value) {
+    this.storageService.set('reverseSearchCoordsFormatEnabled', value);
+    this.igoReverseSearchCoordsFormatEnabled = value;
   }
 
   onSearchSettingsChange() {
@@ -946,7 +906,8 @@ export class PortalComponent implements OnInit, OnDestroy {
   }
 
   searchCoordinate(coord: [number, number]) {
-    this.searchBarTerm = coord.map((c) => c.toFixed(6)).join(', ');
+    this.searchBarTerm = (!this.igoReverseSearchCoordsFormatEnabled) ?
+    coord.map((c) => c.toFixed(6)).join(', ') : coord.reverse().map((c) => c.toFixed(6)).join(', ');
   }
 
   updateMapBrowserClass() {
@@ -1421,8 +1382,9 @@ export class PortalComponent implements OnInit, OnDestroy {
       file,
       features,
       this.map,
+      this.contextState.context$.value.uri,
       this.messageService,
-      this.languageService
+      this.layerService
     );
   }
 
@@ -1430,8 +1392,7 @@ export class PortalComponent implements OnInit, OnDestroy {
     handleFileImportError(
       file,
       error,
-      this.messageService,
-      this.languageService
+      this.messageService
     );
   }
 
