@@ -7,19 +7,17 @@ import {
   CommonVectorStyleOptions
 } from '@igo2/geo';
 
+import { AppOptions, InteractiveTourConfigOptions } from './environnement.interface';
 interface Environment {
   production: boolean;
   igo: {
-    app: {
-      forceCoordsNA: boolean;
-      promotePwa?: boolean;
-    };
+    app: AppOptions,
     catalog?: CatalogServiceOptions;
     importExport?: ImportExportServiceOptions;
     language?: LanguageOptions;
     searchSources?: { [key: string]: SearchSourceOptions };
     projections?: Projection[];
-    interactiveTour?: { tourInMobile: boolean; pathToConfigFile: string };
+    interactiveTour?: InteractiveTourConfigOptions;
     depot?: { url: string; trainingGuides?: string[]; };
     queryOverlayStyle?: {
       base?: CommonVectorStyleOptions,
@@ -38,7 +36,15 @@ export const environment: Environment = {
   production: true,
   igo: {
     app: {
-      forceCoordsNA: false
+      forceCoordsNA: false,
+      install: {
+        enabled: true,
+        promote: true,
+        manifestPath: './config/github.webmanifest'
+      },
+      pwa: {
+        enabled: false,
+      }
     },
     catalog: {
       sources: [
@@ -60,7 +66,9 @@ export const environment: Environment = {
     },
     importExport: {
       url: 'https://geoegl.msp.gouv.qc.ca/apis/ogre',
-      configFileToGeoDBService: './data/geoDataToIDB.json'
+      configFileToGeoDBService: './data/geoDataToIDB.json',
+      clientSideFileSizeMaxMb: 30,
+      allowToStoreLayer: true
     },
     language: {
       prefix: './locale/'
